dispatch all the gfx message function in seperate files
